% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_functions.R
\name{get_account_df}
\alias{get_account_df}
\title{translate account json into df}
\usage{
get_account_df(accounts_json)
}
\arguments{
\item{accounts_json}{json returned from mint api for accounts}
}
\value{
tibble
}
\description{
translate account json into df
}
