% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_functions.R
\name{get_numeric_val_from_config}
\alias{get_numeric_val_from_config}
\title{get_numeric_val_from_config}
\usage{
get_numeric_val_from_config(config_file, field_name)
}
\arguments{
\item{config_file}{user config file}

\item{field_name}{name of numeric field}
}
\value{
numeric
}
\description{
get_numeric_val_from_config
}
