% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_functions.R
\name{get_user_config}
\alias{get_user_config}
\title{obtain user config_file}
\usage{
get_user_config(user_name)
}
\arguments{
\item{user_name}{user name indicating config file}
}
\value{
list
}
\description{
obtain user config_file
}
